refactor: organize file and delete technical detail logic
