+++ src/Routes/router.tsx
@@ -1,5 +1,4 @@
 import AppLayout from "@/layout";
-import DetailTechnician from "@/Pages/detailTechnicianPage";
 import Home from "@/Pages/home";
 import ListClient from "@/Pages/listClient";
 import ListEquipment from "@/Pages/listEquipment";
@@ -59,10 +58,6 @@ const router = createBrowserRouter([
       {
         path: "part",
         element: <ListPart />,
-      },
-      {
-        path: "technician/:id",
-        element: <DetailTechnician />,
       },
       {
         path: "equipment",
